Add test for matching matterns to lua.
I have had use for athis a few times, when dealing with string
represenations of tables which have metatables. While the exact address
changes on each run, we want to be sure that the name of the object is
correct.

Of course, this could be useful for any function which returns
not-exactly-deterministic strings.